seperating the database connection file
