+++ db.js
@@ -1,0 +1,8 @@
+const mysql = require("mysql");
+const db = mysql.createConnection({
+    host:'localhost',
+    user:'root',
+    password:'root',
+    database:'test',
+ });
+ module.exports=db;
